+++ packages/components/collapsible/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @commercetools-uikit/collapsible
 
+## 19.3.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/hooks@19.3.0
+  - @commercetools-uikit/utils@19.3.0
+
 ## 19.2.0
 
 ### Minor Changes