+++ packages/components/field-errors/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @commercetools-uikit/field-errors
 
+## 19.3.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/messages@19.3.0
+
 ## 19.2.0
 
 ### Minor Changes