+++ packages/components/inputs/select-input/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @commercetools-uikit/select-input
 
+## 19.3.0
+
+### Patch Changes
+
+- Updated dependencies [[`91a86d8`](https://github.com/commercetools/ui-kit/commit/91a86d832c89984dd0af786c58d5db478f2f28e1), [`a0626f7`](https://github.com/commercetools/ui-kit/commit/a0626f7de81e3a76bea227bdfb3038ad5339a57a)]:
+  - @commercetools-uikit/select-utils@19.3.0
+  - @commercetools-uikit/design-system@19.3.0
+  - @commercetools-uikit/constraints@19.3.0
+  - @commercetools-uikit/icons@19.3.0
+  - @commercetools-uikit/utils@19.3.0
+
 ## 19.2.0
 
 ### Minor Changes